refactor: 당첨결과를 나타내는 컴포넌트를 모달에서 추출한다
